fix: building

+++ scripts/rebuild.sh
@@ -1,10 +1,14 @@
 #!/bin/bash
 # Remove build directory and rebuild
 cd ..
-rm -rf build/
-mkdir build
-cd build
-cmake ..
-make
-cp ../specs/FIXFXCM10.xml .
-cp ../specs/fxcm.cfg .
+if [ -d build/ ]; then
+	rm -rf build/
+	mkdir build
+fi
+if [ -d build/ ]; then
+	cd build
+	cmake ..
+	make -j4
+#   cp ../specs/FIXFXCM10.xml .
+#	cp ../specs/fxcm.cfg .
+fi
